# trie: Derivesha with stacktrie