--- a/core/types/derive_sha.go
+++ b/core/types/derive_sha.go
@@ -23,7 +23,6 @@ import (
 	"github.com/ethereum/go-ethereum/rlp"
 )
 
-// DerivableList is the interface which can derive the hash.
 type DerivableList interface {
 	Len() int
 	GetRlp(i int) []byte
@@ -39,7 +38,22 @@ type Hasher interface {
 func DeriveSha(list DerivableList, hasher Hasher) common.Hash {
 	hasher.Reset()
 	keybuf := new(bytes.Buffer)
-	for i := 0; i < list.Len(); i++ {
+
+	// StackTrie requires values to be inserted in increasing
+	// hash order, which is not the order that `list` provides
+	// hashes in. This insertion sequence ensures that the
+	// order is correct.
+	for i := 1; i < list.Len() && i <= 0x7f; i++ {
+		keybuf.Reset()
+		rlp.Encode(keybuf, uint(i))
+		hasher.Update(keybuf.Bytes(), list.GetRlp(i))
+	}
+	if list.Len() > 0 {
+		keybuf.Reset()
+		rlp.Encode(keybuf, uint(0))
+		hasher.Update(keybuf.Bytes(), list.GetRlp(0))
+	}
+	for i := 0x80; i < list.Len(); i++ {
 		keybuf.Reset()
 		rlp.Encode(keybuf, uint(i))
 		hasher.Update(keybuf.Bytes(), list.GetRlp(i))