This has the updated version of stacktrie, with `Commit`.