@MariusVanDerWijden 2020-09-02
I think there is something fishy with the shortcut in line 269 as it does not write to the DB.
I created the following test:
```go 
func TestCommit(t *testing.T) {
	db := memorydb.New()
	st := NewStackTrie(db)

	kvs := []struct {
		K string
		V string
	}{
		{K: "405787fa12a823e0f2b7631cc41b3ba8828b3321ca811111fa75cd3aa3bb5ace", V: "1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111"},
	}

	for _, kv := range kvs {
		st.TryUpdate(common.FromHex(kv.K), common.FromHex(kv.V))
	}
	//st.Hash()
	newDB := memorydb.New()

	st.Commit(newDB)
	if db.Len() != newDB.Len() {
		t.Fatalf("Did not insert correctly %v, %v", db.Len(), newDB.Len())
	}

}
```
If I let `st.Hash()` commented out the result is `Did not insert correctly 0, 1` if I delete the comment it results in `Did not insert correctly 1, 0` So it depends on `st.Hash()` into which db we insert.