@holiman 2020-09-02
Added another commit, making it use the existing hasherpool. 
```
[user@work trie]$ benchstat before.txt after.txt 
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-6       579µs ±11%     580µs ±11%     ~     (p=1.000 n=5+5)
DeriveSha200/stack_trie-6     708µs ±18%     624µs ± 5%  -11.90%  (p=0.016 n=5+4)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-6       274kB ± 0%     274kB ± 0%     ~     (p=0.175 n=5+5)
DeriveSha200/stack_trie-6     397kB ± 0%     198kB ± 0%  -50.22%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-6       2.86k ± 0%     2.86k ± 0%     ~     (all equal)
DeriveSha200/stack_trie-6     3.16k ± 0%     2.52k ± 0%  -20.40%  (p=0.008 n=5+5)

```