@holiman 2020-09-02
After this last commit,  It's still a bit slower than `std_trie`, but uses less allocs