@holiman 2020-09-02
On a larger list, with 2K items, it's faster aswell: 
```
BenchmarkDeriveSha2000
BenchmarkDeriveSha2000/std_trie
BenchmarkDeriveSha2000/std_trie-6         	     141	  10510602 ns/op	 4117591 B/op	   33935 allocs/op
BenchmarkDeriveSha2000/stack_trie
BenchmarkDeriveSha2000/stack_trie-6       	     166	   6670803 ns/op	 1958921 B/op	   25090 allocs/op
```