@holiman 2020-09-02
Pushed another fix, now stacktrie beats regular trie on `200` item lists:
```
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-6       585µs ± 8%     577µs ± 2%     ~     (p=0.690 n=5+5)
DeriveSha200/stack_trie-6     804µs ±69%     541µs ± 1%  -32.62%  (p=0.016 n=5+4)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-6       274kB ± 0%     274kB ± 0%     ~     (p=0.175 n=5+5)
DeriveSha200/stack_trie-6     198kB ± 0%     116kB ± 0%  -41.27%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-6       2.86k ± 0%     2.86k ± 0%     ~     (all equal)
DeriveSha200/stack_trie-6     2.52k ± 0%     1.87k ± 0%  -25.63%  (p=0.008 n=5+5)
```