@holiman 2020-09-02
> I think there is something fishy with the shortcut in line 269 as it does not write to the DB.

Yes, it's a bit of an oddity. So the `Hash` _does_ write to database, if non-nil. And the `Commit` just sets the db to non-nil and calls Hash. Maybe it would be neater to change the signature of `hash` into `hash(db)`, and have `Hash()` utilize `hash(nil)`, and have `Commit(db)` use `hash(db)` under the hood?