@holiman 2020-09-02
> So the Hash does write to database, if non-nil. And the Commit just sets the db to non-nil and calls Hash. Maybe it would be neater to change the signature of hash into hash(db), and have Hash() utilize hash(nil), and have Commit(db) use hash(db) under the hood?

It's actually even more complicated than that. Even `Update` (via `insert`) may write to the database. So the `Commit` is just a way to "write leftovers to the db", but in fact db operations happen all the time. Which is totally understandable, since the idea iwth the stacktrie is that we shouldn't build the full trie in memory, but actually get rid of things too. 

And that throws the model of having a totally separate 'Commit`-stage out the window, basically. I'm not sure what the proper thing to do is.