@gballet 2020-09-02
> It's actually even more complicated than that. Even `Update` (via `insert`) may write to the database. So the `Commit` is just a way to "write leftovers to the db", but in fact db operations happen all the time. Which is totally understandable, since the idea iwth the stacktrie is that we shouldn't build the full trie in memory, but actually get rid of things too.

Exactly, you _have_ to commit the nodes as soon as you know they can be hashed, so that you don't run out of memory. We could implement some sort of caching and batch them. I don't think that it is possible to totally separate hashing from committing, though, precisely because memory has to be freed.