@gballet 2020-09-02
rebased to see if this could fix the linter issue, it works fine on my local machine