@holiman 2020-09-02
CPU profile, running the stacktrie benchmark (commenting out the standard trie test)
![profile002](https://user-images.githubusercontent.com/142290/92012000-2870cd00-ed4c-11ea-88af-6c8637ed0115.png)
And mem profile: 
![profile001](https://user-images.githubusercontent.com/142290/92012006-29a1fa00-ed4c-11ea-9b4f-28cad7ea9d05.png)