@holiman 2020-09-03
Before merging this, I think it would be prudent to iterate through the full `ancients` of mainnet, and do the `DeriveSha` for all receipt- and transaction-lists, to verify that it still works. 
As an alternative, I guess a fast-sync would work just as well.