@gballet 2020-09-03
Ok, running a mainnet fastsync :crossed_fingers: