@holiman 2020-09-18
Final numbers: 
```
name                 old time/op    new time/op    delta
DeriveSha200/trie-6     625µs ±20%     515µs ±17%  -17.65%  (p=0.001 n=10+9)

name                 old alloc/op   new alloc/op   delta
DeriveSha200/trie-6     275kB ± 0%      55kB ± 0%  -80.03%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
DeriveSha200/trie-6     2.87k ± 0%     1.27k ± 0%  -55.73%  (p=0.000 n=10+10)
```